Cow Traveling works w/ matrices, but large mazes require too much memory
will try to optimize, both for speed and memory

Will only create a GraphMaze of the portion of the maze that is within
the bounds of T (numMoves/walk length).  Large mazes slow things down
because the adjacency matrix is (mn)^2 * 8 bytes.

In the process of compressing cached adjacency matrix powers in
SparseMatrix if the memory cost gets too high.  Currently works using
HashMaps, but is slow and non-optimal in terms of memory, too.  Will try
another way with my own hash map/trie data structure.